Ajustement des modules pour utiliser le logger
