+++ backend/src/data.mjs
@@ -1,7 +1,6 @@
 import EventEmitter from "node:events";
 import { Buffer } from "node:buffer";
-import chalk from "chalk";
-import logger from "./utils/loggerWinston.mjs";
+import logger from "./utils/logger.mjs";
 
 export default class MyData extends EventEmitter {
     stringDataBuffer = "";
@@ -72,7 +71,7 @@ export default class MyData extends EventEmitter {
                 type: "error",
                 error: "flight mode is unknown (not 0, 1, 2 or 3)",
             });
-            logger.error(`flight mode is unknown (not 0, 1, 2 or 3). Received : ${flightMode}`);
+            logger.warn(`Flight mode is unknown (not 0, 1, 2 or 3). Received : ${flightMode}`, { label: "Data" });
             return;
         }
 
@@ -107,7 +106,9 @@ export default class MyData extends EventEmitter {
                     type: "error",
                     error: `wrong packet length (${line.length} bytes instead of ${prefligthPacketLength})`,
                 });
-                logger.error(`Data : wrong packet length (${line.length} bytes instead of ${prefligthPacketLength})`);
+                logger.warn(`Wrong packet length (${line.length} bytes instead of ${prefligthPacketLength})`, {
+                    label: "Data",
+                });
                 return;
             }
 
@@ -164,7 +165,9 @@ export default class MyData extends EventEmitter {
                     type: "error",
                     error: `wrong packet length (${line.length} bytes instead of ${fligthPacketLength})`,
                 });
-                logger(`Data : wrong packet length (${line.length} bytes instead of ${fligthPacketLength})`);
+                logger.warn(`Wrong packet length (${line.length} bytes instead of ${prefligthPacketLength})`, {
+                    label: "Data",
+                });
                 return;
             }
 
@@ -216,7 +219,9 @@ export default class MyData extends EventEmitter {
                     type: "error",
                     error: `wrong packet length (${line.length} bytes instead of ${postfligthPacketLength})`,
                 });
-                logger.error(`Data : wrong packet length (${line.length} bytes instead of ${postfligthPacketLength})`);
+                logger.warn(`Wrong packet length (${line.length} bytes instead of ${prefligthPacketLength})`, {
+                    label: "Data",
+                });
                 return;
             }
 
@@ -248,7 +253,7 @@ export default class MyData extends EventEmitter {
                 type: "error",
                 error: "cannot parse data",
             });
-            logger.error("cannot parse data");
+            logger.error("Cannot parse data", { label: "Data" });
             return;
         }
