+++ backend/src/webserver.mjs
@@ -23,7 +23,7 @@ export default class MyWebServer {
         // To make the HTTP server of the application listen to client connections
         port = process.env.PORT || port;
         this.server.listen(port, () => {
-            logger(chalk.blue("Web Server"), "listening on port", chalk.yellow(port), `(http://localhost:${port})`);
+            logger.info(`Listening on port ${chalk.blue(port)} (http://localhost:${port})`, { label: "WebServer" });
         });
     }
